* commit '4f2c1a1825e333d9308001ca02805f34c4a0175a':
  Fix 32-bit build.

+++ src/core/testing/plg_python27_unit_test/reflection_test_module.cpp
@@ -1693,7 +1693,7 @@ void checkDict( const Collection & dictResult,
 	// Note that Python dictionaries are unordered
 	// So it needs to lookup by key so that the expected value is known
 	CHECK_EQUAL( expectedSize, dictResult.size() );
-	for (int i = 0; i < expectedSize; ++i)
+	for (int i = 0; i < static_cast< int >( expectedSize ); ++i)
 	{
 		sprintf( buffer, "%d", i );
 		bufferStr = buffer;
